+++ CHANGELOG.md
@@ -1,15 +1,15 @@
 # 2.8.6
 
 Changes:
-* Due to newly introduced configs, JSONC option `{ "temperatureUnit": "C" }` has been changed to `{ "temperature": { "unit": "C" } }`
+* Due to newly introduced configs, JSONC option `{ "temperatureUnit": "C" }` has been changed to `{ "temp": { "unit": "C" } }`
 
 Bugfixes:
 * Fix incorrect GPU name detection for Intel iGPU on Linux (#736, GPU, Linux)
 
 Features:
 * Support additional temperature formatting options (#737)
-    * `{ "temperature": { "ndigits": 1 } }`
-    * `{ "temperature": { "color": { "green": "green", "yellow": "yellow", "red": "red" } } }`
+    * `{ "temp": { "ndigits": 1 } }`
+    * `{ "temp": { "color": { "green": "green", "yellow": "yellow", "red": "red" } } }`
 
 # 2.8.5